feat(mrml-core): Add `Fragment` element
This component effectively just acts as a collection of children.
However, when rendered, they behave as if directly under the
parent element. This is useful for passing around reusable
components, etc, in Rust code.

In this commit, it is not actually possible to parse a
Fragment. Instead, they must be constructed by Rust code.
This is because of some non-configurable validation in
xmlparser:
https://github.com/RazrFalcon/xmlparser/blob/e54c2768f20814140c11e6c92f94ee74bfd54808/src/stream.rs#L432

Signed-off-by: Jade Ellis <jade@ellis.link>